#129, refactor project structure
