+++ src/main/java/org/artembogomolova/demo/webapp/dao/repo/IAuthorityRepository.java
@@ -1,6 +1,6 @@
-package org.artembogomolova.demo.webapp.dao;
+package org.artembogomolova.demo.webapp.dao.repo;
 
-import org.artembogomolova.demo.webapp.model.auth.Authority;
+import org.artembogomolova.demo.webapp.domain.auth.Authority;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
