+++ src/main/java/org/artembogomolova/demo/webapp/dao/repo/ICategoryRepository.java
@@ -1,6 +1,6 @@
-package org.artembogomolova.demo.webapp.dao;
+package org.artembogomolova.demo.webapp.dao.repo;
 
-import org.artembogomolova.demo.webapp.model.business.Category;
+import org.artembogomolova.demo.webapp.domain.business.Category;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
