+++ src/main/java/org/artembogomolova/demo/webapp/dao/repo/IOrderPositionRepository.java
@@ -1,6 +1,6 @@
-package org.artembogomolova.demo.webapp.dao;
+package org.artembogomolova.demo.webapp.dao.repo;
 
-import org.artembogomolova.demo.webapp.model.business.OrderPosition;
+import org.artembogomolova.demo.webapp.domain.business.OrderPosition;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
