+++ src/main/java/org/artembogomolova/demo/webapp/dao/repo/IOrderRepository.java
@@ -1,6 +1,6 @@
-package org.artembogomolova.demo.webapp.dao;
+package org.artembogomolova.demo.webapp.dao.repo;
 
-import org.artembogomolova.demo.webapp.model.business.Order;
+import org.artembogomolova.demo.webapp.domain.business.Order;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
