+++ src/main/java/org/artembogomolova/demo/webapp/dao/repo/IPhysicalAddressRepository.java
@@ -1,6 +1,6 @@
-package org.artembogomolova.demo.webapp.dao;
+package org.artembogomolova.demo.webapp.dao.repo;
 
-import org.artembogomolova.demo.webapp.model.core.PhysicalAddress;
+import org.artembogomolova.demo.webapp.domain.core.PhysicalAddress;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
