+++ src/main/java/org/artembogomolova/demo/webapp/dao/repo/ITicketRepository.java
@@ -1,6 +1,6 @@
-package org.artembogomolova.demo.webapp.dao;
+package org.artembogomolova.demo.webapp.dao.repo;
 
-import org.artembogomolova.demo.webapp.model.business.Ticket;
+import org.artembogomolova.demo.webapp.domain.business.Ticket;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
