+++ src/main/java/org/artembogomolova/demo/webapp/dao/repo/IUserRoleRepository.java
@@ -1,10 +1,11 @@
-package org.artembogomolova.demo.webapp.dao;
+package org.artembogomolova.demo.webapp.dao.repo;
 
-import org.artembogomolova.demo.webapp.model.auth.Role;
+import org.artembogomolova.demo.webapp.domain.auth.Role;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface IUserRoleRepository extends CrudRepository<Role, Long> {
 
+  Role findByName(String name);
 }
