+++ src/main/java/org/artembogomolova/demo/webapp/domain/auth/BasicAuthorityEnum.java
@@ -1,7 +1,7 @@
-package org.artembogomolova.demo.webapp.model.auth;
+package org.artembogomolova.demo.webapp.domain.auth;
 
 import java.util.Locale;
-import org.artembogomolova.demo.webapp.model.core.IdentifiedEntity;
+import org.artembogomolova.demo.webapp.domain.core.IdentifiedEntity;
 
 public enum BasicAuthorityEnum {
   BAT_CREATE("create"),
