+++ src/main/java/org/artembogomolova/demo/webapp/domain/business/Order.java
@@ -1,4 +1,4 @@
-package org.artembogomolova.demo.webapp.model.business;
+package org.artembogomolova.demo.webapp.domain.business;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -17,9 +17,9 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
-import org.artembogomolova.demo.webapp.model.core.IdentifiedEntity;
-import org.artembogomolova.demo.webapp.model.core.Person;
-import org.artembogomolova.demo.webapp.model.core.PhysicalAddress;
+import org.artembogomolova.demo.webapp.domain.core.IdentifiedEntity;
+import org.artembogomolova.demo.webapp.domain.core.Person;
+import org.artembogomolova.demo.webapp.domain.core.PhysicalAddress;
 
 @Entity
 @Table(name = "orders")
