+++ src/main/java/org/artembogomolova/demo/webapp/domain/business/Producer.java
@@ -1,4 +1,4 @@
-package org.artembogomolova.demo.webapp.model.business;
+package org.artembogomolova.demo.webapp.domain.business;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,8 +9,8 @@ import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import lombok.Data;
-import org.artembogomolova.demo.webapp.model.core.IdentifiedEntity;
-import org.artembogomolova.demo.webapp.model.core.PhysicalAddress;
+import org.artembogomolova.demo.webapp.domain.core.IdentifiedEntity;
+import org.artembogomolova.demo.webapp.domain.core.PhysicalAddress;
 
 @Entity
 @Table(name = "producers")
