+++ src/test/java/org/artembogomolova/demo/webapp/dao/CategoryRepositoryTest.java
@@ -2,7 +2,8 @@ package org.artembogomolova.demo.webapp.dao;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.artembogomolova.demo.webapp.model.business.Category;
+import org.artembogomolova.demo.webapp.dao.repo.ICategoryRepository;
+import org.artembogomolova.demo.webapp.domain.business.Category;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
