+++ src/test/java/org/artembogomolova/demo/webapp/dao/ProducerRepositoryTest.java
@@ -2,7 +2,8 @@ package org.artembogomolova.demo.webapp.dao;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.artembogomolova.demo.webapp.model.business.Producer;
+import org.artembogomolova.demo.webapp.dao.repo.IProducerRepository;
+import org.artembogomolova.demo.webapp.domain.business.Producer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.repository.CrudRepository;
